feat(pep621): use class based processors
