+++ lib/modules/manager/pep621/processors/index.ts
@@ -1,0 +1,3 @@
+import { PdmProcessor } from './pdm';
+
+export const processors = [new PdmProcessor()];
